pyproject.toml
[project]
name = "garagem"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "BlackVSK", email = "phrod2007@gmail.com"},
]
dependencies = ["django>=5.2.1", "djangorestframework>=3.16.0"]
requires-python = "==3.13.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
